First

+++ src/main/java/com/iqmsoft/dao/InventoryRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.iqmsoft.dao;
 
 public interface InventoryRepositoryCustom {
 
-    int updateStockProduct(String id, int stock);
+    long updateStockProduct(String id, int stock);
 
-    int updatePriceProduct(String id, double price);
+    long updatePriceProduct(String id, double price);
 }
